recipes-support: ptest-runner: Add missing ptest-runner-collect-system-data
Add a bbappend to do the same as upstream openembedded-core commit
4e6be3fb521b23cfc175d0c09725bcc3ebbc73b2.

When ptest-runner times out or otherwise fails, it tries to
call ptest-runner-collect-system-data, so install the script.

The script currently calls dmesg, df, free (which are provided
by busybox, etc.) and pstree (which is a sub-package of psmisc).

Add pstree as an RDEPENDS.

Signed-off-by: Tim Orling <timothy.t.orling@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
[Converted to bbappend for poky/hardknott]
Signed-off-by: Chris Paterson <chris.paterson2@renesas.com>